Split langium into `node` and `browser` package
