+++ packages/langium/node.d.ts
@@ -1,0 +1,2 @@
+/* eslint-disable header/header */
+export * from './lib/node';
