+++ packages/langium/package.json
@@ -19,6 +19,10 @@
   "files": [
     "lib",
     "src",
+    "browser.js",
+    "browser.d.ts",
+    "node.js",
+    "node.d.ts",
     "test.js",
     "test.d.ts"
   ],
