+++ packages/langium/src/browser/index.ts
@@ -1,0 +1,10 @@
+/******************************************************************************
+ * Copyright 2022 TypeFox GmbH
+ * This program and the accompanying materials are made available under the
+ * terms of the MIT License, which is available in the project root.
+ ******************************************************************************/
+
+export * from './browser-module';
+
+// Re-export everything from the main index
+export * from '../index';
