+++ packages/langium/src/node/index.ts
@@ -1,0 +1,11 @@
+/******************************************************************************
+ * Copyright 2022 TypeFox GmbH
+ * This program and the accompanying materials are made available under the
+ * terms of the MIT License, which is available in the project root.
+ ******************************************************************************/
+
+export * from './node-file-system-provider';
+export * from './node-module';
+
+// Re-export everything from the main index
+export * from '../index';
