+++ packages/langium/test/validation/grammar-validation.test.ts
@@ -4,7 +4,7 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import { createLangiumGrammarServices } from '../../src';
+import { createLangiumGrammarServices } from '../../src/node';
 import { Assignment, CrossReference, Grammar, Group, ParserRule } from '../../src/grammar/generated/ast';
 import { IssueCodes } from '../../src/grammar/langium-grammar-validator';
 import { expectError, expectNoIssues, expectWarning, validationHelper, ValidationResult } from '../../src/test';
